% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING_UTILITIES.R
\name{AIC.tvRRR}
\alias{AIC.tvRRR}
\title{Corrected version (for small samples) taken from
Burnham, Anderson (2004): Multimodel Inference: Understanding AIC and BIC in model selection
Recommended if t / K < 40, but as it converges to zero with growing t it should be used anyway
for that correction is set as default}
\usage{
\method{AIC}{tvRRR}(kf, d, model = "A", correct = TRUE)
}
\description{
Corrected version (for small samples) taken from
Burnham, Anderson (2004): Multimodel Inference: Understanding AIC and BIC in model selection
Recommended if t / K < 40, but as it converges to zero with growing t it should be used anyway
for that correction is set as default
}
