% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY.R
\name{approx.kronecker}
\alias{approx.kronecker}
\title{Approximates kronecker product decomposition. (Daniel Kapla)}
\usage{
approx.kronecker(C, dimA, dimB)
}
\arguments{
\item{C}{desired kronecker product result.}

\item{dimA}{length 2 vector of dimensions of \code{A}.}

\item{dimB}{length 2 vector of dimensions of \code{B}.}
}
\value{
list with attributes \code{A} and \code{B}.
}
\description{
Approximates the matrices \code{A} and \code{B} such that
C = A \%x\% B
with \code{\%x\%} the kronecker product of the matrixes \code{A} and \code{B}
of dimensions \code{dimA} and \code{dimB} respectively.
}
\examples{
A <- matrix(seq(14), 7, 2)
B <- matrix(c(TRUE, FALSE), 3, 4)
C <- kronecker(A, B) # the same as 'C <- A \%x\% B'
approx.kronecker(C, dim(A), dim(B))

}
\seealso{
C.F. Van Loan / Journal of Computational and Applied Mathematics
123 (2000) 85-100 (pp. 93-95)
}
\keyword{internal}
