% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{dataset}
\alias{dataset}
\title{DATA GENERATING FUNCTIONS:
This needs to be documented very carefully}
\usage{
dataset(type, p, d, q, t, X, forecast, model = "A", ...)
}
\arguments{
\item{type}{\code{= c("deterministic", "break", "VARbreak", "rw", "fixed")},
see Details for a description of the different kinds}

\item{p}{dimension of the target vector}

\item{d}{latent dimension of the model}

\item{q}{dimension of the predictors}

\item{t}{length of the time series}

\item{X}{optional, the predictors used to generate the time series.
Not used if \code{type = "VARbreak"}. If not handed over, it is drawn
randomly from a q-variate normal distribution with mean zero and
covariance identity}

\item{forecast}{how many additional variables do we want to generate for
forecasting experiments}

\item{model}{which model do we want? Either "A" or "B"}

\item{alpha_0}{optional}

\item{alpha_1}{optional (only used if \code{model = "A"} and
\code{type = c("break", "VARbreak", "deterministic")} )}

\item{beta_0}{optional}

\item{beta_1}{optional (only used if \code{model = "B"} and
\code{type = c("break", "VARbreak", "deterministic")} )}

\item{Omega}{error covariance, defaults to the identity matrix}

\item{shift}{only used if \code{type = c("break", "VARbreak", "deterministic")}),
guides how far the matrices are apart (see Details)}

\item{Sigma}{column covariance if \code{type = "rw"}}

\item{Delta}{row covariance if \code{type = "rw"}, defaults to the identity}

\item{breakpoint}{position of the structural break in the dataset for
\code{type = c("break", "VARbreak")}}
}
\value{
A named list with the desired dataset, containing components
\item{y}{the target variable (t x p) matrix}
\item{X}{the predictors (t x q) matrix}
\item{alpha}{if \code{model = "A"}, an array of dimensions (t, p, d),
if \code{model = "B"}, a (p, d) matrix}
\item{beta}{if \code{model = "B"}, an array of dimensions (t, q, d),
if \code{model = "A"}, a (q, d) matrix}
\item{Omega}{the error covariance (p, p) dimensional}
\item{Sigma}{if \code{type = "rw"} the column covariance of the states, (d, d) matrix}
}
\description{
DATA GENERATING FUNCTIONS:
This needs to be documented very carefully
}
\details{

}
