% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING.R
\name{eval_tvRRR}
\alias{eval_tvRRR}
\title{Filtering:}
\usage{
eval_tvRRR(X, y, u = NULL, d, model = "A", alpha, beta, ...)
}
\arguments{
\item{X}{the predictors}

\item{y}{the target variable}

\item{u}{additional predictors that come in as full rank}

\item{d}{latent dimension (min. 1, no default)}

\item{model}{specifies whether we're filtering in model "A" or model "B"}

\item{alpha}{starting value for the algorithm, or the time-constant parameter matrix (p x d)}

\item{beta}{starting value for the algorithm, or time-constant parameter matrix (q x d)}

\item{Gamma}{(optional), the time constant full rank (t x k) matrix}

\item{P_00}{starting covariance for the algorithm (p\emph{d x p}d matrix for A, (q\emph{d x q}d) matrix for B)}

\item{Sigma}{column covariance of the states alpha_t (symmetric d x d matrix)}

\item{Omega}{error covariance in the measurement equation (symmetric p x p matrix)}
}
\value{
An object of class \code{tvRRR} (LINK TO DOCUMENTATION OF tvRRR CLASS ->
DAS IST EIGENTLICH GELOGEN, mal gucken wie wir das lösen können)
}
\description{
Run the Kalman filter for a fixed set of parameters and prespecified model
}
