% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING.R
\name{filter_modelA}
\alias{filter_modelA}
\title{Function that runs the Kalman filter}
\usage{
filter_modelA(
  y,
  X,
  u = NULL,
  beta,
  alpha_00,
  Gamma = NULL,
  P_00,
  Sigma,
  Omega,
  d,
  return_covariances = TRUE,
  Gamma_rrr = "identity"
)
}
\arguments{
\item{y}{the target variable (t x p matrix)}

\item{X}{the predictors (t x q matrix)}

\item{u}{(optional) additional predictors that do not necessarily vary in
time (t x k matrix)}

\item{beta}{the time-constant part of the matrix (either NULL or a starting value)}

\item{alpha_00}{starting value for the algorithm (either NULL or a starting value)}

\item{P_00}{starting covariance for the algorithm (p\emph{d x p}d matrix)}

\item{Sigma}{column covariance of the states alpha_t (symmetric d x d matrix)}

\item{Omega}{error covariance in the measurement equation (symmetric p x p matrix)}

\item{d}{latent dimension (min. 1, no default)

Output:
A named list of lists with elements
\itemize{
\item states: filtered (the filtered states) -- one state matrix per row (t + 1 x p * d)
smoothed (the smoothed states) -- one state matrix per row (t + 1 x p * d)
\item covariances: the filtered and smoothed covariances and lag-1 covariances
\code{P_t^t} filtered covariances -- array of dimensions (t+1, p\emph{d, p}d)
\code{P_t^t-1}predicted covariances -- (t, p\emph{d, p}d)
\code{P_t^T} smoothed covariances -- (t+1, p\emph{d, p}d)
\verb{P_t-1t-2^T} smoothed lag-1 covariances -- (t, p\emph{d, p}d)
\item data: the data handed over to the algorithms
\code{X} predictors -- (t, q)
\code{y} responses -- (t, p)
\code{u} additional predictors -- (t, k)
\code{Z} transition matrices (X_t'beta (x) I_p) -- (t, p, p*d)
\item parameters used during filtering: Sigma (d, d), Omega (p, p), beta (q, d)
}

Symmetry of the matrices is enforced using the hint in
https://math.stackexchange.com/questions/2335831/covariance-matrix-p-for-an-extended-kalman-filter-not-symmetric}

\item{gamma}{(optional), the time constant full rank (t x k) matrix}
}
\description{
Input:
}
