% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING.R
\name{filter_modelB}
\alias{filter_modelB}
\title{Function that runs the Kalman filter for model (B)}
\usage{
filter_modelB(
  X,
  y,
  u = NULL,
  P_00,
  Sigma,
  Omega,
  beta_00,
  alpha,
  Gamma = NULL,
  d,
  return_covariances = TRUE,
  Gamma_rrr = "identity"
)
}
\description{
It processes beta', not beta!!
}
