% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING.R
\name{fit_modelA}
\alias{fit_modelA}
\title{Runs the expectation maximization algorithm for Model A}
\usage{
fit_modelA(
  X,
  y,
  u = NULL,
  d,
  beta = NULL,
  alpha_00 = NULL,
  Gamma = NULL,
  P_00 = 1000 * diag(ncol(y) * d),
  Sigma = NULL,
  Omega = NULL,
  Omega_diagonal = FALSE,
  maxit = 100,
  silent = FALSE,
  tol_finish = 0.001,
  tol_EMstep = 0.001,
  return_covariances = FALSE,
  Gamma_rrr = "identity",
  initialize = "RRR"
)
}
\arguments{
\item{X}{predictors (t x q-dimensional)}

\item{y}{target (t x p-dimensional)}

\item{u}{additional predictors (t x q-dimensional)}

\item{d}{latent dimension}

\item{beta}{starting value for beta, default NULL (RRR)}

\item{alpha_00}{starting value for the algorithm, default NULL (RRR)}

\item{Gamma}{starting value for the fixed full-rank coefficient matrix,
default NULL (RRR)}

\item{P_00}{starting state covariance (default 1000 * diag(p x d))}

\item{Sigma}{column covariance of states (default 0.01 * diag(d))}

\item{Omega}{error covariance (defaults to residual covariance from RRR,
or respective starting values)}

\item{Omega_diagonal}{logical, indicates whether Omega is assumed to be a
diagonal matrix (advisable if p is large)}

\item{maxit}{maximum number of iterations for the EM algorithm}

\item{silent}{logical, indicates whether progress should be printed during
model fitting}

\item{tol_finish}{tolerance for stopping the EM algorithm}

\item{tol_EMstep}{tolerance for iterative estimation during EM step

Output:}

\item{return_covariances}{logical, indicates whether the state covariances
should be returned (might be necessary for evaluation
of the likelihood).}

\item{Gamma_rrr}{type of normalization for the starting values obtained from RRR}

\item{initialize}{either \code{"RRR"} or \code{"random"}, applies if no starting
values are handed over}
}
\value{
An object of class \code{tvRRR}, that is a named list of lists with elements
\item{states}{ The estimated states (i.e. coefficient matrices)
\itemize{
\item filtered (the filtered states)
\item smoothed (the smoothed states)
}
}
\item{covariances}{The filtered and smoothed covariances and lag-1 covariances
(if \code{return_covariances = TRUE})
\itemize{
\item \code{P_t^t} filtered covariances
\item \code{P_t^t-1} predicted covariances
\item \code{P_t^T} smoothed covariances
\item \code{P_t-1t-2^T} smoothed lag-1 covariances
}
}
\item{data}{the data handed over to the algorithms
\itemize{
\item \code{X} predictors
\item \code{y} responses
\item \code{u} additional predictors
\item \code{Z} transition matrices (X_t'beta (x) I_p)
}
}
\item{parameters}{The parameters that have been fitted during the algorithm,
that is\itemize{
\item \code{Sigma} the column covariance of the states
\item \code{Omega} the error covariance
\item \code{beta} (for model A)
\item \code{alpha} (for model B)
}
}
\item{likelihoods}{list containing Q and data loglikelihood for each iteration}
\item{convergence_information}{Message containing parameter stages at convergence}
}
\description{
Runs the expectation maximization algorithm for Model A
}
