% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING.R
\name{fit_tvRRR}
\alias{fit_tvRRR}
\title{Fit the tvRRR model (Brune et al. (2021+)) using the EM algorithm:}
\usage{
fit_tvRRR(X, y, u = NULL, d, model = "A", ...)
}
\arguments{
\item{X}{predictors (t x q-dimensional numeric matrix)}

\item{y}{target (t x p-dimensional numeric matrix)}

\item{u}{additional predictors (t x k-dimensional numeric matrix)}

\item{d}{latent dimension}

\item{...}{additional parameters that are handed over to \code{fit_modelA()} or
\code{fit_modelB()} respectively (starting values, convergence conditions, etc.)}
}
\value{
An object of class \code{tvRRR}, that is a named list of lists with elements
\item{states}{ The estimated states (i.e. coefficient matrices)
\itemize{
\item filtered (the filtered states)
\item smoothed (the smoothed states)
}
}
\item{covariances}{The filtered and smoothed covariances and lag-1 covariances
(if \code{return_covariances = TRUE})
\itemize{
\item \code{P_t^t} filtered covariances
\item \code{P_t^t-1} predicted covariances
\item \code{P_t^T} smoothed covariances
\item \code{P_t-1t-2^T} smoothed lag-1 covariances
}
}
\item{data}{the data handed over to the algorithms
\itemize{
\item \code{X} predictors
\item \code{y} responses
\item \code{u} additional predictors
\item \code{Z} transition matrices (X_t'beta (x) I_p)
}
}
\item{parameters}{The parameters that have been fitted during the algorithm,
that is\itemize{
\item \code{Sigma} the column covariance of the states
\item \code{Omega} the error covariance
\item \code{beta} (for model A)
\item \code{alpha} (for model B)
}
}
\item{likelihoods}{list containing information on the data likelihood during
the fitting procedure, i.e. \itemize{
\item \code{Q} the negative expected log likelihood obtained from EM
\item \code{logLik} actual data log Likelihood for each iteration
}
}
\item{convergence_information}{Message containing parameter stages at convergence}
}
\description{
Fit the tvRRR model (Brune et al. (2021+)) using the EM algorithm:
}
\details{
The \code{tvRRR()} function fits the tvRRR models of class A and B. The model
is a reduced rank regression model, where the coefficient matrix can be decomposed
in two ways: C_t = alpha_t beta' (Model A) or C_t = alpha beta_t' (Model B). For
details on the model fitting algorithm see Brune et al. (2021+).

\code{print()} and \code{fitted()} methods are provided.
}
\examples{
\dontrun{
X <- mvtnorm::rmvnorm(50, rep(0, 5), diag(5))
y <- mvtnorm::rmvnorm(50, rep(0, 5), diag(5))

mod <- tvRRR(X = X, y = y, d = 2)

print(mod)
}

}
\seealso{
\code{\link[tvRRR]{fit_modelA}}, \code{\link[tvRRR]{fit_modelB}}
}
\keyword{internal}
