% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING_UTILITIES.R
\name{fitted.tvRRR}
\alias{fitted.tvRRR}
\title{Calculate fitted values from tvRRR object}
\usage{
\method{fitted}{tvRRR}(object, type = "filtered", model, ...)
}
\arguments{
\item{object}{an object of class \code{tvRRR}}

\item{type}{either \code{filtered} or \code{smoothed}, defaults to filtered}

\item{model}{\code{"A"} or \code{"B"}, can be handed over, but otherwise is determined
automatically from the \code{tvRRR} object}

\item{...}{ignored}
}
\description{
Calculate fitted values from tvRRR object
}
