% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY.R
\name{grassmann_dist}
\alias{grassmann_dist}
\title{UNNORMALIZED GRASSMANN DISTANCE (BASED ON PRINCIPAL ANGLES)}
\usage{
grassmann_dist(
  A,
  B,
  tol = max(dim(A)) * sqrt(.Machine$double.eps),
  normalized = FALSE
)
}
\description{
UNNORMALIZED GRASSMANN DISTANCE (BASED ON PRINCIPAL ANGLES)
}
\keyword{internal}
