% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{make_dataset_break}
\alias{make_dataset_break}
\title{Dataset that contains a structural break}
\usage{
make_dataset_break(
  X,
  alpha_0 = NULL,
  alpha_1 = NULL,
  shift = 1,
  beta_0 = NULL,
  beta_1 = NULL,
  Omega = NULL,
  p,
  d,
  model = "A",
  forecast = 0,
  breakpoint = floor(nrow(X)/2),
  ...
)
}
\description{
Dataset that contains a structural break
}
\keyword{internal}
