% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{make_dataset_deterministic}
\alias{make_dataset_deterministic}
\title{Dataset that has a deterministic transition}
\usage{
make_dataset_deterministic(
  X,
  alpha_0 = NULL,
  alpha_1 = NULL,
  shift = 1,
  beta_0 = NULL,
  beta_1 = NULL,
  Omega = NULL,
  p,
  d,
  model = "A",
  forecast = 0,
  ...
)
}
\description{
Dataset that has a deterministic transition
}
\keyword{internal}
