% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{make_dataset_fixed}
\alias{make_dataset_fixed}
\title{Matrices are actually fixed}
\usage{
make_dataset_fixed(
  X,
  alpha = NULL,
  beta = NULL,
  Omega = NULL,
  p,
  d,
  model = "A",
  forecast = 0
)
}
\description{
Matrices are actually fixed
}
\keyword{internal}
