% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{make_dataset_rw}
\alias{make_dataset_rw}
\title{Dataset where the time-varying parameter matrices perform a random walk in
euclidean space where Cov(vec(alpha_t)) = kronecker(Sigma, Delta), Sigma
is dxd (column covariance) and Delta is pxp (row covariance)}
\usage{
make_dataset_rw(
  X,
  alpha_0 = NULL,
  beta_0 = NULL,
  Omega = NULL,
  p,
  d,
  q = ncol(X),
  Sigma = NULL,
  Delta = if (model == "A") diag(p) else diag(q),
  model = "A",
  forecast = 0,
  ...
)
}
\description{
Dataset where the time-varying parameter matrices perform a random walk in
euclidean space where Cov(vec(alpha_t)) = kronecker(Sigma, Delta), Sigma
is dxd (column covariance) and Delta is pxp (row covariance)
}
