% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{make_VARdataset_break}
\alias{make_VARdataset_break}
\title{VAR-dataset that contains a break}
\usage{
make_VARdataset_break(
  t = 100,
  alpha_0 = NULL,
  alpha_1 = NULL,
  beta_0 = NULL,
  beta_1 = NULL,
  Omega = NULL,
  p,
  d,
  model = "A",
  shift = 1,
  forecast = 0,
  lag = 1,
  breakpoint = floor(t/2)
)
}
\description{
VAR-dataset that contains a break
}
\keyword{internal}
