% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\name{make_VARdataset_determ}
\alias{make_VARdataset_determ}
\title{VAR-dataset with deterministic transition}
\usage{
make_VARdataset_determ(
  t = 100,
  alpha_0 = NULL,
  alpha_1 = NULL,
  beta_0 = NULL,
  beta_1 = NULL,
  Omega = NULL,
  p,
  d,
  model = "A",
  shift = 1,
  forecast = 0,
  lag = 1,
  breakpoint = floor(t/2)
)
}
\description{
VAR-dataset with deterministic transition
}
\keyword{internal}
