% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY.R
\name{matpow}
\alias{matpow}
\title{GENERALIZED MATRIX POWER FUNCTION FOR SYMMETRIC MATRICES (by Daniel Kapla)}
\usage{
matpow(A, pow, symmetric = FALSE, tol = 1e-07)
}
\description{
GENERALIZED MATRIX POWER FUNCTION FOR SYMMETRIC MATRICES (by Daniel Kapla)
}
\keyword{internal}
