% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY.R
\name{polar_proj}
\alias{polar_proj}
\title{POLAR PROJECTION OF A MATRIX ONTO ITS ORTHOGONAL PART}
\usage{
polar_proj(X)
}
\description{
POLAR PROJECTION OF A MATRIX ONTO ITS ORTHOGONAL PART
}
\keyword{internal}
