% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING_UTILITIES.R
\name{print.tvRRR}
\alias{print.tvRRR}
\title{Print Method for tvRRR object}
\usage{
\method{print}{tvRRR}(x, ...)
}
\arguments{
\item{x}{an object of class \code{tvRRR}}

\item{...}{ignored}
}
\value{
A short summary of the \code{tvRRR} object.
}
\description{
Print Method for tvRRR object
}
