% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY.R
\name{rr.regression}
\alias{rr.regression}
\title{RRR with additional full-rank predictors
taken from Reinsel and Velu (1998), Ch. 3, Thm. 3.1
u corresponds to their z}
\usage{
rr.regression(X, y, u = NULL, rank, Gamma_type = "identity")
}
\description{
Model: y = mu + ABX + Du + error
}
