% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY.R
\name{subsp_dist}
\alias{subsp_dist}
\title{NORMALIZED SUBSPACE DISTANCE (FROBENIUS OF PROJECTION DIFFERENCE)}
\usage{
subsp_dist(X, Y)
}
\description{
NORMALIZED SUBSPACE DISTANCE (FROBENIUS OF PROJECTION DIFFERENCE)
}
\keyword{internal}
