% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETER_UPDATES.R
\name{X_calc_Q}
\alias{X_calc_Q}
\title{HELP 3: Calculate the current value of the Likelihood and (if required) the
matrix C for parameter estimation}
\usage{
X_calc_Q(kf, p, d, q, t, return_C = FALSE, model = "A")
}
\description{
HELP 3: Calculate the current value of the Likelihood and (if required) the
matrix C for parameter estimation
}
\keyword{internal}
