% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETER_UPDATES.R
\name{X_eval_lik}
\alias{X_eval_lik}
\title{HEPL 3.1:
Calculate the true data likelihood based on the latent states}
\usage{
X_eval_lik(kf, Omega, beta = NULL, alpha = NULL, d, y, X)
}
\description{
HEPL 3.1:
Calculate the true data likelihood based on the latent states
}
\keyword{internal}
