% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETER_UPDATES.R
\name{X_update_alpha}
\alias{X_update_alpha}
\title{HELP 1B: update alpha during the EM algorithm}
\usage{
X_update_alpha(d, Omega, Gamma = NULL, X, y, u = NULL, s, kf, p, q, t)
}
\description{
HELP 1B: update alpha during the EM algorithm
}
\keyword{internal}
