% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETER_UPDATES.R
\name{X_update_beta}
\alias{X_update_beta}
\title{HELP 1A: update beta during the EM algorithm}
\usage{
X_update_beta(d, Omega, Gamma, X, y, u, kf, p, q, t)
}
\description{
HELP 1A: update beta during the EM algorithm
}
\keyword{internal}
