% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETER_UPDATES.R
\name{X_update_Omega}
\alias{X_update_Omega}
\title{HELP 2: update Omega during the EM algorithm, optionally including the external variables}
\usage{
X_update_Omega(Z, Gamma, Omega_diagonal, y, u, kf, p, t)
}
\description{
HELP 2: update Omega during the EM algorithm, optionally including the external variables
}
\keyword{internal}
