% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING_UTILITIES.R
\name{fitted.tvRRR}
\alias{fitted.tvRRR}
\title{Calculate fitted values from tvRRR object}
\usage{
\method{fitted}{tvRRR}(kf, type = "filtered", model)
}
\arguments{
\item{kf}{an object of class \code{tvRRR}}

\item{type}{either \code{filtered} or \code{smoothed}, defaults to filtered}

\item{model}{\code{"A"} or \code{"B"}, can be handed over, but is determined
automatically from the tvRRR object}
}
\description{
Calculate fitted values from tvRRR object
}
