% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING_UTILITIES.R
\name{predict.tvRRR}
\alias{predict.tvRRR}
\title{Predict method for tvRRR object}
\usage{
\method{predict}{tvRRR}(kf, newdata = NULL)
}
\arguments{
\item{kf}{object of class tvRRR as returned by \code{\link[tvRRR]{tvRRR}()} function}

\item{newdata}{can be specified differently with differing behavior: \itemize{
\item if \code{NULL}, the \code{\link[tvRRR]{fitted.tvRRR}()} method is called
\item a named list of length 1 or 2 containing predictors X and u for predictions
using the last state fitted
\item a named list of length 2 or 3 containing targets y, predictors X and, optionally
additional predictors u
}}
}
\value{
The predicted target variables in the specified setup (as a matrix)
}
\description{
Predict method for tvRRR object
}
