% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITTING.R
\name{tvRRR}
\alias{tvRRR}
\title{Fit the tvRRR model (Brune et al., 2021+) to an observed dataset.}
\usage{
tvRRR(
  X,
  y,
  u = NULL,
  model = "A",
  d,
  select_rank = TRUE,
  d_max,
  criterion = "BIC",
  correctAIC = TRUE,
  Sigma_init = 0.1,
  ...
)
}
\arguments{
\item{X}{predictors (t x q-dimensional numeric matrix)}

\item{y}{target (t x p-dimensional numeric matrix)}

\item{u}{additional predictors (t x k-dimensional numeric matrix)}

\item{d}{latent dimension}

\item{d_max}{maximum latent dimension, needs to be an integer and can at most be min(p, q)}

\item{criterion}{string specifying the desired criterion, in \code{c("AIC", "BIC")},
defaults to AIC}

\item{correctAIC}{logical, if criterion is AIC, determines the need to correct for
small sample sizes, defaults to TRUE}

\item{...}{additional parameters that are handed over to \code{fit_modelA()} or
\code{fit_modelB()} respectively (starting values, convergence conditions, etc.)}
}
\value{
An object of class \code{tvRRR}, that is a named list of lists with elements
\item{states}{ The estimated states (i.e. coefficient matrices)
\itemize{
\item filtered (the filtered states)
\item smoothed (the smoothed states)
}
}
\item{covariances}{The filtered and smoothed covariances and lag-1 covariances
(if \code{return_covariances = TRUE})
\itemize{
\item \code{P_t^t} filtered covariances
\item \code{P_t^t-1} predicted covariances
\item \code{P_t^T} smoothed covariances
\item \code{P_t-1t-2^T} smoothed lag-1 covariances
}
}
\item{data}{the data handed over to the algorithms
\itemize{
\item \code{X} predictors
\item \code{y} responses
\item \code{u} additional predictors
\item \code{Z} transition matrices (X_t'beta (x) I_p)
}
}
\item{parameters}{The parameters that have been fitted during the algorithm,
that is\itemize{
\item \code{Sigma} the column covariance of the states
\item \code{Omega} the error covariance
\item \code{beta} (for model A)
\item \code{alpha} (for model B)
}
}
\item{likelihoods}{list containing information on the data likelihood during
the fitting procedure, i.e. \itemize{
\item \code{Q} the negative expected log likelihood obtained from EM
\item \code{logLik} actual data log Likelihood for each iteration
}
}
\item{convergence_information}{Message containing parameter stages at convergence}
\item{AIC (or BIC)}{Value of the information criterion for the selected model}
}
\description{
Rank selection
is performed automatically based on either AIC or BIC. The model with the lowest
AIC / BIC is selected. We use a version of AIC that is corrected for small
sample sizes (AIC_c).
}
\seealso{
\code{\link[tvRRR]{fit_modelA}}, \code{\link[tvRRR]{fit_modelB}}
}
