% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETER_UPDATES.R
\name{update_pars}
\alias{update_pars}
\title{Perform one EM-step}
\usage{
update_pars(
  kf,
  p,
  d,
  t,
  q,
  tol = 0.001,
  maxit = 50,
  Omega_diagonal = FALSE,
  model = "A"
)
}
\arguments{
\item{kf}{a Kalman filter object, output from eval_tvRRR() / filter_modelA() / filter_modelB() function}

\item{p}{dimension of the target}

\item{d}{latent dimension}

\item{t}{length of the time series}

\item{q}{dimension of the predictors}

\item{tol}{tolerance for convergence in iterative parameter updating}

\item{Omega_diagonal}{logical, indicates whether Omega is assumed to be diagonal}
}
\value{
List of updated parameters:
\item{Omega}{}
\item{beta}{(when fitting model A)}
\item{alpha}{(when fitting model B)}
\item{Sigma}{}
\item{Gamma}{(when considering external variables)}
\item{Q}{value of the expected likelihood Q(Theta|Theta^j)}
}
\description{
Function that updates the parameters from the current output of the KF-function
i.e. one updating step in the EM algorithm
}
\keyword{internal}
